+++ src/schemaGenerator.ts
@@ -25,6 +25,7 @@ import {
   GraphQLType,
   GraphQLInterfaceType,
   GraphQLFieldMap,
+  GraphQLUnionType,
 } from 'graphql';
 
 import {
@@ -386,7 +387,10 @@ function addResolveFunctionsToSchema(
   resolveFunctions: IResolvers,
   resolverValidationOptions: IResolverValidationOptions = {},
 ) {
-  const { allowResolversNotInSchema = false } = resolverValidationOptions;
+  const {
+    allowResolversNotInSchema = false,
+    requireResolversForResolveType = false,
+  } = resolverValidationOptions;
 
   Object.keys(resolveFunctions).forEach(typeName => {
     const type = schema.getType(typeName);
@@ -403,7 +407,6 @@ function addResolveFunctionsToSchema(
     Object.keys(resolveFunctions[typeName]).forEach(fieldName => {
       if (fieldName.startsWith('__')) {
         // this is for isTypeOf and resolveType and all the other stuff.
-        // TODO require resolveType for unions and interfaces.
         type[fieldName.substring(2)] = resolveFunctions[typeName][fieldName];
         return;
       }
@@ -459,6 +462,24 @@ function addResolveFunctionsToSchema(
         setFieldProperties(field, fieldResolve);
       }
     });
+  });
+
+  // If we have any union or interface types throw if no there is no resolveType or isTypeOf resolvers
+  if (requireResolversForResolveType) {
+    checkForResolveTypeResolver(schema);
+  }
+}
+
+function checkForResolveTypeResolver(schema: GraphQLSchema) {
+  Object.keys(schema.getTypeMap())
+  .map(typeName => schema.getType(typeName))
+  .forEach((type: GraphQLUnionType | GraphQLInterfaceType) => {
+    if (!(type instanceof GraphQLUnionType || type instanceof GraphQLInterfaceType)) {
+      return;
+    }
+    if (!type.resolveType) {
+      throw new SchemaError(`Type ${type.name} is missing a "resolveType" resolver`);
+    }
   });
 }